test: 네트워크 테스트 XCTest → Swift Testing 마이그레이션
- NetworkClient/PhotoRepository/KeychainTokenStore 테스트를 @Suite + @Test(한글 설명) + #expect로 전환
- StubURLProtocol static 상태를 공유하는 스위트는 @Suite(.serialized) 컨테이너(StubURLProtocolSuites)로 직렬화 — Swift Testing의 스위트 간 병렬 실행 경합 방지
- NetworkError에 Equatable 추가 → #expect(throws:)로 에러 값 비교
- onRefreshFailed 콜백 검증을 CallbackFlag 헬퍼 대신 confirmation()으로 대체
- URLProtocol 큐에서의 assert를 recordedRequests 사후 검증으로 재구성 (테스트 태스크 밖 #expect 기록 문제 회피)
- 성능 테스트는 measure API 대응이 없어 XCTest 유지
