+++ Rephoto_iOSTests/Support/TestHelpers.swift
@@ -6,7 +6,17 @@
 //
 
 import Foundation
+import Testing
 @testable import Rephoto_iOS
+
+// MARK: - 직렬화 컨테이너
+
+/// StubURLProtocol의 전역 상태(handler/recordedRequests)를 공유하는 스위트들의 컨테이너.
+///
+/// Swift Testing은 스위트 간에도 병렬 실행하므로, 스텁 상태를 공유하는 스위트는
+/// 이 컨테이너의 extension으로 선언해 전체를 직렬 실행시킨다 (.serialized는 자식에 재귀 적용).
+@Suite(.serialized)
+enum StubURLProtocolSuites {}
 
 // MARK: - StubURLProtocol
